Ref issue #14 Edit a lot
Fix the capitalization.

Format all occurances of paths to use inline literal "`path`".

Remove the "remove CNAME" section.  If a contributor changes the "CNAME" file and includes it in a pull request then the pull request cannot be merged directly.

Make the section level consistent (four dial symbols instead of five for the level inside the one of three dials).